created azure function transforming data and storing data to cosmos db
