+++ azuredemo/EventHubSender/src/main/java/com/knoldus/Main.java
@@ -1,4 +1,6 @@
-package org.example;
+package com.knoldus;
+
+import com.knoldus.Publisher.Sender;
 
 public class Main {
     public static void main(String[] args) {
